[IDLE-467] 이전 메시지를 읽고 있을 때 새로운 메시지가 왔을 경우 최 하단으로 가지 않도록 수정
