+++ feature/center/applicant-inquiry/src/main/java/com/tgyuu/applicant/inquiry/ApplicantInquiryFragment.kt
@@ -70,7 +70,7 @@ internal class ApplicantInquiryFragment : BaseComposeFragment() {
                     applicants = applicants,
                     navigateTo = {
                         navigationHelper.navigateTo(
-                            com.idle.navigation.NavigationEvent.NavigateTo(
+                            com.idle.navigation.NavigationEvent.To(
                                 it
                             )
                         )
