+++ feature/center/chatting/src/main/java/com/idle/center/chatting/CenterChattingFragment.kt
@@ -67,7 +67,7 @@ internal class CenterChattingFragment : BaseComposeFragment() {
                 chatRoomList = chatRoomList,
                 navigateTo = {
                     navigationHelper.navigateTo(
-                        com.idle.navigation.NavigationEvent.NavigateTo(
+                        com.idle.navigation.NavigationEvent.To(
                             it
                         )
                     )
