+++ feature/chatting-detail/src/main/java/com/idle/chatting_detail/ChattingDetailFragment.kt
@@ -30,7 +30,6 @@ import androidx.compose.ui.unit.dp
 import androidx.fragment.app.viewModels
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import androidx.navigation.fragment.findNavController
-import androidx.navigation.fragment.navArgs
 import com.idle.chatting_detail.component.CareChatReceiverTextBubble
 import com.idle.chatting_detail.component.CareChatSenderTextBubble
 import com.idle.chatting_detail.component.CareChatSenderTextBubbleWithImage
@@ -45,31 +44,29 @@ import com.idle.domain.model.chat.ChatMessage
 import com.idle.domain.model.profile.CenterProfile
 import com.idle.domain.model.profile.WorkerProfile
 import com.idle.domain.util.formatYearMonthDate
+import com.idle.navigation.NavigationEvent
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
 internal class ChattingDetailFragment : BaseComposeFragment() {
-    private val args: ChattingDetailFragmentArgs by navArgs()
     override val fragmentViewModel: ChattingDetailViewModel by viewModels()
 
     @Composable
     override fun ComposeLayout() {
-        val chattingRoomId = args.chattingRoomId
-        val receiverUserType = UserType.create(args.receiverUserType)
-        val receiverId = args.receiverId
-        val senderId = args.senderId
-
         fragmentViewModel.apply {
+            val receiverUserType = receiverUserType
+            val receiverId = receiverId
+
             val writingText by writingText.collectAsStateWithLifecycle()
             val chatMessages by chatMessages.collectAsStateWithLifecycle()
             val workerProfile by workerProfile.collectAsStateWithLifecycle()
             val centerProfile by centerProfile.collectAsStateWithLifecycle()
 
             LaunchedEffect(Unit) {
-                getUserProfile(myUserType = receiverUserType, senderId = senderId)
-                getChatMessages(myUserType = receiverUserType, chattingRoomId)
-                subscribeChatMessage(receiverId)
-                readMessage(roomId = chattingRoomId, myUserType = receiverUserType)
+                getUserProfile()
+                getChatMessages()
+                subscribeChatMessage()
+                readMessage()
             }
 
             if (chatMessages != null && workerProfile != null && centerProfile != null) {
@@ -81,219 +78,13 @@ internal class ChattingDetailFragment : BaseComposeFragment() {
                     writingText = writingText,
                     chatMessages = chatMessages!!,
                     onWritingTextChange = ::setWritingText,
-                    getChatMessages = {
-                        getChatMessages(
-                            myUserType = receiverUserType,
-                            chattingRoomId
-                        )
-                    },
-                    sendMessage = {
-                        sendMessage(
-                            myUserType = receiverUserType,
-                            roomId = chattingRoomId
-                        )
-                    },
-                    navigateTo = {
-                        navigationHelper.navigateTo(
-                            com.idle.navigation.NavigationEvent.NavigateTo(
-                                it
-                            )
-                        )
-                    },
+                    getChatMessages = ::getChatMessages,
+                    sendMessage = ::sendMessage,
+                    navigateTo = { navigationHelper.navigateTo(NavigationEvent.To(it)) },
                     navigateUp = { findNavController().navigateUp() }
                 )
             } else {
                 ChattingDetailLoadingScreen(navigateUp = { findNavController().navigateUp() })
-            }
-        }
-    }
-}
-
-@Composable
-internal fun ChattingDetailScreen(
-    receiverId: String,
-    myUserType: UserType,
-    workerProfile: WorkerProfile,
-    centerProfile: CenterProfile,
-    writingText: String,
-    chatMessages: List<ChatMessage>,
-    onWritingTextChange: (String) -> Unit,
-    sendMessage: () -> Unit,
-    getChatMessages: () -> Unit,
-    navigateTo: (com.idle.navigation.DeepLinkDestination) -> Unit,
-    navigateUp: () -> Unit,
-) {
-    val focusManager = LocalFocusManager.current
-    val listState = rememberLazyListState()
-    var lastDate: String? = null
-
-    val lastVisibleIndex by remember {
-        derivedStateOf {
-            listState.firstVisibleItemIndex + listState.layoutInfo.visibleItemsInfo.size - 1
-        }
-    }
-
-    val isNearEnd = lastVisibleIndex >= chatMessages.size.minus(3)
-
-    LaunchedEffect(isNearEnd) {
-        if (chatMessages.isNotEmpty() && isNearEnd) {
-            getChatMessages()
-        }
-    }
-
-    Scaffold(
-        topBar = {
-            val title = when (myUserType) {
-                UserType.WORKER -> centerProfile.centerName
-                UserType.CENTER -> workerProfile.workerName
-            }
-
-            CareSubtitleTopBar(
-                title = title,
-                onNavigationClick = navigateUp,
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(start = 12.dp, top = 48.dp, end = 20.dp, bottom = 12.dp),
-            )
-        },
-        containerColor = CareTheme.colors.white000,
-        modifier = Modifier.addFocusCleaner(focusManager),
-    ) { paddingValue ->
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(paddingValue),
-        ) {
-            LazyColumn(
-                state = listState,
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .weight(1f)
-                    .background(CareTheme.colors.gray050)
-                    .padding(horizontal = 18.dp),
-            ) {
-                item {
-                    Spacer(
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .padding(20.dp)
-                    )
-                }
-
-                itemsIndexed(
-                    items = chatMessages,
-                    key = { _, item -> item.id },
-                ) { index, chatMessage ->
-                    val isLast = if (index < chatMessages.size - 1) {
-                        val nextIndex = index + 1
-                        chatMessage.senderId != chatMessages[nextIndex].senderId
-                    } else {
-                        true
-                    }
-                    val padding = PaddingValues(bottom = if (isLast) 16.dp else 6.dp)
-
-                    val messageDate = chatMessage.createdAt.formatYearMonthDate()
-                    val showDate = lastDate != messageDate
-                    if (showDate) {
-                        lastDate = messageDate
-
-                        Text(
-                            text = messageDate,
-                            style = CareTheme.typography.caption1,
-                            color = CareTheme.colors.gray700,
-                            textAlign = TextAlign.Center,
-                            modifier = Modifier
-                                .fillMaxWidth()
-                                .padding(
-                                    top = if (index == 0 || isLast) 0.dp else 10.dp,
-                                    bottom = 16.dp
-                                )
-                        )
-                    }
-
-                    val isMyMessage = receiverId == chatMessage.senderId
-                    if (isMyMessage) {
-                        CareChatReceiverTextBubble(
-                            chatMessage = chatMessage,
-                            isLast = isLast,
-                            onSeeAllChatClicked = {
-                                navigateTo(
-                                    com.idle.navigation.DeepLinkDestination.SeeAllChat(
-                                        it
-                                    )
-                                )
-                            },
-                            modifier = Modifier.padding(padding),
-                        )
-                    } else {
-                        val showProfile = if (index > 0) {
-                            val previousIndex = index - 1
-                            chatMessage.senderId != chatMessages[previousIndex].senderId
-                        } else {
-                            true
-                        }
-
-                        if (showProfile) {
-                            CareChatSenderTextBubbleWithImage(
-                                imageUrl = when (myUserType) {
-                                    UserType.CENTER -> workerProfile.profileImageUrl
-                                    UserType.WORKER -> centerProfile.profileImageUrl
-                                },
-                                senderName = when (myUserType) {
-                                    UserType.CENTER -> workerProfile.workerName
-                                    UserType.WORKER -> centerProfile.centerName
-                                },
-                                chatMessage = chatMessage,
-                                isLast = isLast,
-                                onSeeAllChatClicked = {
-                                    navigateTo(
-                                        com.idle.navigation.DeepLinkDestination.SeeAllChat(
-                                            it
-                                        )
-                                    )
-                                },
-                                modifier = Modifier.padding(padding),
-                            )
-                        } else {
-                            CareChatSenderTextBubble(
-                                chatMessage = chatMessage,
-                                isLast = isLast,
-                                isRead = isLast,
-                                onSeeAllChatClicked = {
-                                    navigateTo(
-                                        com.idle.navigation.DeepLinkDestination.SeeAllChat(
-                                            it
-                                        )
-                                    )
-                                },
-                                modifier = Modifier.padding(padding),
-                            )
-                        }
-                    }
-                }
-            }
-
-            Row(
-                verticalAlignment = Alignment.CenterVertically,
-                horizontalArrangement = Arrangement.spacedBy(16.dp),
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(start = 20.dp, end = 20.dp, top = 8.dp, bottom = 16.dp)
-            ) {
-                CareChatTextField(
-                    value = writingText,
-                    onValueChanged = onWritingTextChange,
-                    hint = "메세지를 입력하세요.",
-                    modifier = Modifier.weight(1f),
-                )
-
-                Image(
-                    painter = painterResource(com.idle.designresource.R.drawable.ic_send_message),
-                    contentDescription = "",
-                    modifier = Modifier
-                        .size(32.dp)
-                        .clickable(throttleTime = 1000L) { sendMessage() },
-                )
             }
         }
     }
