+++ feature/chatting-detail/src/main/java/com/idle/see_all_chat/SeeAllChatViewModel.kt
@@ -1,4 +1,4 @@
-package com.idle.chatting_detail.seeallchat
+package com.idle.see_all_chat
 
 import androidx.lifecycle.ViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
