+++ feature/chatting-detail/src/main/res/navigation/nav_chatting_detail.xml
@@ -24,7 +24,7 @@
     </fragment>
     <fragment
         android:id="@+id/seeAllChatFragment"
-        android:name="com.idle.chatting_detail.seeallchat.SeeAllChatFragment"
+        android:name="com.idle.see_all_chat.SeeAllChatFragment"
         android:label="SeeAllChatFragment">
         <deepLink app:uri="@string/chatting_see_all_chat_deeplink_url" />
         <argument
