+++ feature/job-posting-detail/src/main/java/com/idle/worker/job/posting/detail/center/CenterJobPostingDetailFragement.kt
@@ -147,7 +147,7 @@ internal class CenterJobPostingDetailFragment : BaseComposeFragment() {
                                 deleteJobPosting = ::deleteJobPosting,
                                 navigateTo = {
                                     navigationHelper.navigateTo(
-                                        com.idle.navigation.NavigationEvent.NavigateTo(
+                                        com.idle.navigation.NavigationEvent.To(
                                             it
                                         )
                                     )
