+++ feature/worker/chatting/src/main/java/com/idle/worker/chatting/WorkerChattingFragment.kt
@@ -67,7 +67,7 @@ internal class WorkerChattingFragment : BaseComposeFragment() {
                     chatRoomList = chatRoomList,
                     navigateTo = {
                         navigationHelper.navigateTo(
-                            com.idle.navigation.NavigationEvent.NavigateTo(
+                            com.idle.navigation.NavigationEvent.To(
                                 it
                             )
                         )
