+++ feature/worker/home/src/main/java/com/idle/worker/home/WorkerHomeFragment.kt
@@ -93,7 +93,7 @@ internal class WorkerHomeFragment : BaseComposeFragment() {
                 removeFavoriteJobPosting = ::removeFavoriteJobPosting,
                 navigateTo = {
                     navigationHelper.navigateTo(
-                        com.idle.navigation.NavigationEvent.NavigateTo(
+                        com.idle.navigation.NavigationEvent.To(
                             it
                         )
                     )
